Show treebark photos from FireBase
Show treebark photos from FireBase in Instagram style.
Investigated a bit about using Netlilfy to avoid exposing FireBase API key, but decided that exposing FireBase API key does not harm that much that it worth the hassle of moving the whole site from Jekyll to some dynamic model. But Netlify seems to be a nice tool to try out in the future.